Fixing pointer and scroll events for nested iframes
